version 1.0.3

+++ src/main/kotlin/eu/ibagroup/r2z/UssFile.kt
@@ -1,3 +1,13 @@
+/*
+ * This program and the accompanying materials are made available under the terms of the
+ * Eclipse Public License v2.0 which accompanies this distribution, and is available at
+ * https://www.eclipse.org/legal/epl-v20.html
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Copyright © 2021 IBA Group, a.s.
+ */
+
 package eu.ibagroup.r2z
 
 import com.google.gson.annotations.Expose
@@ -15,11 +25,11 @@ data class UssFile(
 
   @SerializedName("size")
   @Expose
-  val size: Int? = null,
+  val size: Long? = null,
 
   @SerializedName("uid")
   @Expose
-  val uid: Int? = null,
+  val uid: Long? = null,
 
   @SerializedName("user")
   @Expose
@@ -27,7 +37,7 @@ data class UssFile(
 
   @SerializedName("gid")
   @Expose
-  val gid: Int? = null,
+  val gid: Long? = null,
 
   @SerializedName("group")
   @Expose
